Experimental modification that transforms C++ pass-by-value to references. Specifically, a function like this:
   void blah(Foo f);

gets wrapped like it was declared as follows if 'Foo" was declared as a class:

   void blah(Foo &f);

This mod attempts to fix the horrible problem of wrappers for pass-by-value
with classes that lack default constructors.

This modification might break some typemap code--especially if it was
written for pass-by-value (it will no longer be matched).

I'm not sure if this is the best solution, but it does make the problem go away.


git-svn-id: https://swig.svn.sourceforge.net/svnroot/swig/branches/rel-1-3@2229 626c5289-ae23-0410-ae9c-e8d60b6d4f22