Changed the way that constants are wrapped to work for all constants. The value is now taken from C code. Previously the constant was set as Java code and worked for most, but not all cases.
Dave,
This is a rather a big hack. Can you do a diff on the change to have a look to see how you can incorporate a neater change into the core? Probably something in Swig_VargetToFunction(). Maybe some typemaps should be used for determining the code that emit_action() generates?

Also, while implementing this I wanted to get a debug dump of all the attributes in the functionWrapper node as they are different to what -dump_tree generates. Is there a way to do this yet?

Thanks


git-svn-id: https://swig.svn.sourceforge.net/svnroot/swig/branches/rel-1-3@2764 626c5289-ae23-0410-ae9c-e8d60b6d4f22