All but two test-suite tests succeed:
funcptr fails: the type is bunched in front of the variable name, so
that it doesn't parse right in C.  The type of the function pointer
needs to wrap.  I will ask Mr. Beazley how this is done.

template_ns4: I don't even understand what this code does.  I need
to ask for help on it.


git-svn-id: https://swig.svn.sourceforge.net/svnroot/swig/branches/rel-1-3@3027 626c5289-ae23-0410-ae9c-e8d60b6d4f22