Add %feature("java:abstractdirector").
The abstract directors feature will on demand turn C++ classes with
pure virtual functions into abstract java classes with abstract
methods. The feature is controlled with:

%abstractdirector
%noabstractdirector
%clearabstractdirector

and works either globally or on a class or on a specific
method. Globally it will affect all pure virtual methods, given a
class it will just affect that class, and given a method just that
specific method (although it will of course have to make the class
abstract).