Basic support for C# on MonoTouch / Xamarin.iOS
Add in attributes to ensure the Mono ahead of time compiler does not
remove reverse callbacks / delegates for calling managed code from
managed code. See
http://docs.xamarin.com/guides/ios/advanced_topics/limitations#Reverse_Callbacks

The Xamarin.iOS MonoTouch.MonoNativeFunctionWrapper and
MonoTouch.MonoPInvokeCallback attributes need not be defined and used
in the MonoTouch assembly in order for them to be used, so we put
them in the more convenient intermediary PINVOKE class.