Fix ccache-swig internal error bug due to premature file cleanup
Fixes SF bug 1319 which shows up as a failure in the ccache tests on
Debian 64 bit Wheezy, possibly because ENABLE_ZLIB is defined.

This bug is due to files being too aggressively cleaned up part way
through the caching. The .stderr file is cached and then retrieved
from the cache for displaying to stderr. However, the stats are updated
between caching and using the .stderr file. During the stats update the
cache is cleaned and the newly cached files can be removed if the max
number of files per directory is low. Really the cache should be cleaned
up at exit to solve this (as is done in ccache-3.1). The workaround fix
ensures the cached files are ignored during cleanup, which is a bit
tricky as sometimes files from a previous run have the same time stamp,
so that don't appear to be the oldest in the cache.