Fix test suite for languages without Doxygen support.
"-doxygen" can't be used in SWIGOPT for the languages other than Java and
Python, this results in immediate error as it's not supported by SWIG when
using any other language.

Use this switch -- and run the Doxygen tests -- only for the languages that do
support it and, to avoid slowing down the test suite even more, only for the
tests that need it.

Also allow running only the Doxygen tests using check-doxygen target.