Describe Doxygen-specific %features in a single place.
Put them all in a single table instead of describing doxygen:notranslate
separately for Java and Python when it does the same thing for both.

And while doxygen:nolinkranslate and doxygen:nostripparams are currently
Java-specific, it probably won't remain this way and these features would
either be made to work for Python as well or, perhaps, removed entirely.