Fix handling of non-first arguments default values in Python.
Since 15b3690, non-first arguments with default values of non-primitive types
were not used in Python wrappers, resulting in invalid Python code if the
preceding arguments did have default values.

This happened because is_representable_as_pyargs() didn't check any arguments
except the first one at all in most cases because it used tmap:in:next
unconditionally, but it is usually null. Fall back on the next argument in the
function parameter list now to fix this.