[Python] Fix handling of default numeric arguments values.
Instead of just checking that the first character is a digit and blithely
assuming that the rest must be a Python number, really parse the number and
check that it's indeed representable in Python.

This gets rid of false positives (e.g. octal numbers using C syntax are not
valid in Python 3; long long interger literals are not valid in any version of
Python at all) but also of false negatives (e.g. negative numbers such as even
a simple -1 were deemed to be never representable in Python).