[Python] Don't use non-global C++ enums as default values.
This doesn't work for enums inherited from the base class as they need to be
explicitly qualified in Python, so avoid using any non-global enum elements at
all in Python. This is not ideal as in many places we could still do it, but
at least it doesn't result in invalid Python code.

Closes #296.