[MATLAB] Introduce the possibility to have define setup script
The setup script can be defined as:

%matlabsetup %{

matlab code

%}

For a project foo, this code will end up in a 'foosetup.m'.

The code is automatically run, just before the generated mex file is
called for the first time.

The code runs only once unless matlab memory of global variables is
cleared.